Add new command buildWDKAutoSwitch
This command executes a predefined wdk test pipeline.
As buildWDK it tests the package agains the configured number
of unity version.

This new version sets environment variables to switch unity versions
with the help of . The project
has to use this gradle plugin to be able to switch/install the correct version.