Add buildJavaLibraryOSSRH (#5)
Description
===========

I need to add a new buildJavaLibrary step which
can publish to OSSRH. The old step is using bintray
credentials and can't really be reused or extended. For
the future 2.x version of the pipelines I plan to fix the
naming schemes and maybe move the publish configuration
out of the general pipeline.

I also fixed a runtime bug in this pipeline. With OSSRH it is
possible to publish also snapshot builds. The try..catch setup
was broken and never discovered as most pipelines skip the
publish step on normal ci runs and skip the check stage on publish.
The pipeline still executed the publish stage even though the check
stage failed. I removed the catch statement which just printed the
error message. Now a failing tests stopps the pipeline.

I also adjusted the TestHelper class to allow null or empty values
for the platform parameter.

Changes
=======

* ![ADD] new step `buildJavaLibraryOSSRH`
* ![FIX] `publish` stage execution after failing `checks` stage
* ![IMPROVE] `Testhelperi#transformIntoCheckStep` make `platform` optional