Improve Java OSSRH pipeline with sonarqube publish and optional ws clean
Description
===========

This patch adds some improvements and fixes from the build gradle plugin
pipeline. I fixed the default values for platform in the same way we
fixed it for buildGradlePlugin in commit d88125bc.
I also copied the setup for the sonarqube publish.
As a last improvement I also added a new parameter `CLEAN_WORKSPACE`
for both `buildGradlePlugin`, `buildJavaLibraryOSSRH` and `buildWDKAutoSwitch`

The idea is to make the cleanup controlable during invocation of
the pipeline. The default value is `true`.

Changes
=======

* ![FIX] `config.platform` default value lookup
* ![ADD] sonarqube publish setup in `buildJavaLibraryOSSRH`
* ![IMPROVE] cleanup workspace with pipeline parameter