Fix local integration tests
Description
===========

This patch sets the grgit credentials needed to run the tests
locally before the tests are executed. On jenkins it is done
withing the pipeline which was actually not needed.

There was also an invocation to `gitHttpTransportUrl` which
is both a getter and a method. But it seems that in a newer version
of the github API the method is gone.

Changes
=======

* ![FIX] local integration tests.