Improve internal plugin setup
Description
===========

This patch switches the `buildSrc` plugin setup
to a gradle composite build. This allows us to split
the ci tests for the actual content of this repo and
the build logic into two different CI calls.

Some scripts and the new internal Jenkinsfile are contain
temp additions. We need a new custom plugin of our internal
gradle development plugin to support private plugins.

Changes
=======

* ![IMPROVE] ![GRADLE] internal plugin setup.