Remove sonarqube branch limitation and added branch name property (#41)
Sonarqube will now execute for all branches, in order to provider PR information to sonarqube. Sonarqube differentiates between PR analysis and plain branch analysis, and give precedence to later. So, if the branch property is set, sonarqube will run the analysis as plain branch one.  

Thus, we force this property to empty on PR analysis in order to sonarqube to detect this run as a PR analysis. On non-PR runs, the branch name property is set to `$BRANCH_NAME`.

PS: right now, this is only working 100% for the `buildGradlePlugin` pipeline. Other pipelines gradle plugins does not support the sonarqube task yet, so please do not enable sonarqube coverage by setting a value to `config.sonarToken` until support is provided.


Co-authored-by: Joaquim Neto <joaquim.neto@wooga.net>