Refactor pipeline scripts and added tests (#34)
Long-waited pipeline tests refactoring! Added tests for WDK and java/gradle pipelines. The pipeline file tests are based in the DeclarativeJenkinsSpec class, which has been expanded now with support to many more jenkins functionalities, and a fake secrets storage and environment. Pipeline tests are inside the scripts folder, with check step tests isolated into the JavaCheckSpec and WDKCheckSpec classes to avoid repetition.

On the code itself, the main entrypoints are the Checks, Setups, Assemblers and Publishers classes, which contains methods for dealing with their respective operation. gradleWrapper.groovy script behavior is now isolated in the Gradle class, upon which most of the application depends upon. In order to make things more "dsl-like" functionalities from these classes were wrapped on scripts, with the javaLibCheck.groovy script containing the check operation for all java libraries/gradle plugins, and the wdkOps.groovy containing the setup, assembly and check operation related to the WDK pipeline. Finally publish operations are in the publish.groovy script.

The configurations were also turned into classes, in order to improve reuse and reliability. The Config class and its related classes are configuration for java pipelines, and the WDKConfig class is configuration for WDK pipelines. There are many intersections between those configurations, most of them belonging into the Platform class. I believe that there is a way to merge these configurations further, but to avoid breaking changes and to avoid coupling too much the WDK and the Java pipelines, it was preferred to leave these two independent of each other by now.

Co-authored-by: Joaquim Neto <joaquim.neto@wooga.net>