Fix gradleWrapper calling error
Description
===========

This patch tries to fix a calling error at runtime from the game
projects.

> No signature of method: java.lang.Class.fromJenkins() is applicable for argument types: (gradleWrapper, java.lang.String, null) values: [gradleWrapper@7274fcad, info, null]

It looks like the issue is with the check
`params.STACK_TRACE as boolean`

I tried to setup a simple testcase for this but it never fails.
My susspission is that jenkins is doing some transformations
differently then what we have in our testing harness.

I will test this patch on production builds in the hope
that it fixes the issue.

I switched from `boolean` to `Boolean` as it is the default for groovy
and box types can be null which should fix this already.
On top I also added a null check. This might be overkill because of
the first change.

Changes
=======

* ![FIX] `gradleWrapper` calling error
* ![CHANGE] `boolean` to `Boolean` in `Gradle.fromJenkins`
* ![ADD] `null` check in `gradleWrapper` for `params.STACK_TRACE`