Remove personal git credentials for user Larusso
Description
===========

This change replaces the usage of an github credentials
pair which uses a user token ([larusso]).
I could have changed the underlying credential in AWS and
it would have switched over automatically. But I have two reasons
for it. First I don't like the credentials id `github_up`. It was
meant as a push credential. I think that 'github_access`
is clearer. Second I still don't know if the new token works in
all scenarios. The `github_up` token is also used in the game pipelines.
Here I would like to maybe get away from a shared token. But all
in all I don't 100% know if this new token just works or if
we break things.

Changes
=======

* ![IMPROVE] replace personal git credentials with shared user