added refresh dependencies parameter working for all pipelines (#49)
Added working refresh dependencies parameter to all pipelines and to gradle wrapper.

Please note that in the way that things are right now, every gradle call will trigger the --refresh-dependencies behavior. I'm not sure if this is what we want, as it may slow down some things, but we can talk about it ;)

Some minor code refactors are here as well, as removing the legacy RUN_SONAR parameter, and improvments in the GradleWrapperSpec tests.