Sandboxed tests (#52)
## Description
 Test now run in a sandbox similar to the one in Jenkins itself. The `GroovySandbox` class is provided by the `script-security` jenkins plugin, there was need of heavy configuration to make it work properly. 
 
 In order for the sandbox to work, it needs to be in a groovy sandboxed environment, that should include a groovy loader (`GroovyScriptEngine` or `GroovyShell`) with:

 * A "GroovySandbox Secure class loader", and
 * A GroovySandbox Secure compiler configurations". 
 
 These configurations are merged to the existing ones in our test library in the `SandboxPipelineTestHelper` class. 

The sandbox receives whitelists that indicates which elements will be allowed on it. By default, everything is denied by the sandbox, so every function that you want to allow needs to be whitelisted. The `tools.sandbox.PackageWhitelist` whitelists everything inside a package, and the `MethodSignatureWhitelist` whitelists all method signatures passed to it.  We also use some jenkins-made whtelists as the `ClassLoaderWhitelist` and the `GenericWhitelist` classes.

Finally, you can use the sandbox inside a test by the `loadSandboxedScript` and `inSandbox` methods, which loads a script using the sandbox and runs a closure inside the sandbox respectively. Please be aware that non-basic JVM objects can't be passed from the 'normal' environment to the sandboxed one, as they have different classloaders. More details in the docs of the `DeclarativeJenkinsSpec.inSandbox` method.


Co-authored-by: Joaquim Neto <joaquim.neto@wooga.net>