Fix workspace stash pattern for buildWDKAutoSwitch
## Description
I had an issue in one WDK that the `paket.lock` file was missing. That is actually never an issue because it simply gets regenerated. But in my case a tool tried to access it before it got regenerated. I saw now that the pattern had a bug and the file actually never got picked up.

## Changes
* ![FIX] workspace stash pattern for buildWDKAutoSwitch