Fixed dockerargs object argument strings and improved tests (#114)
## Description
Arguments generation for Docker projects with docker containers are broken, as build arguments would enter the real argument list as an array, instead of a string with space separated arguments. Tests were testing the wrong string as well, so they couldn't catch the error. Added another layer of unit tests for the DockerArgs object and the jenkins-level tests now tests the right behavior.

Branch based on `fix/late-checkout` (#105)

## Changes
* ![FIX] argument generation bug for Docker projects