[ATLAS-831] Adds buildLocalGradlePlugin pipeline using inner plugin as sample (#129)
Adds new pipeline for local plugin project tests, vars/buildLocalGradlePlugin.groovy. Its set to be triggered whenever a change in a given checkDirectory is passed through, and then runs in the same framework as the other java libraries (vars/javaLibs.groovy), except it never publishes anything. If checkDirectory is empty it will assume that the git root is the location of the local project.

This new pipeline was tested and is applied in this repository local plugin atlas-jenkins-pipeline-gradle. This PR also updates the local plugin build.gradle file to use the lastest version of atlas-plugins.