Add RW credentials for release builds, add missing properties to pipeline [ATLAS-1559][ATLAS-1566] (#258)
## Description
The publish process for new WDKs is broken as it skips the github release step which also creates the needed version tag in the project. This yields errors the next time a new release is issued. Worse even the old release can be overridden by the next publish. To combat this this PR patches both issues at once. First part is the quick addition of the old legacy properties `release.stage` and `release.scope` These should not be needed anymore but it seems that in some
WDKs the usage is still required. Its easier to fix this here instead of pushing updates through gradle.

The second issue is the publish user. For snapshot builds we need the ability to override versions on jfrog. But that capability is very dangerous for rc or final builds. So we now switch the publish user credentials based on the build release type.

## Changes
* ![ADD] Credentials for final releases
* ![FIX] Missing Properties for backwards compatibility

Co-authored-by: Robin Hentschel <flakysalt@users.noreply.github.com>